+++ .mulle/etc/env/environment-user-nat.sh
@@ -4,3 +4,15 @@
 export MULLE_CRAFT_POST_PROJECT="/home/nat/.mulle/bin/reamalgamate-mulle-core"
 
 
+#
+#
+#
+export MULLE_CRAFT_CCACHE="sccache"
+
+
+#
+#
+#
+export MULLE_SDE_EDITOR_CHOICE="subl"
+
+