+++ src/mulle-buffer.c
@@ -45,9 +45,6 @@ struct mulle_buffer   *mulle_buffer_create( struct mulle_allocator *allocator)
 {
    struct mulle_buffer  *buffer;
 
-   if( ! allocator)
-      allocator = &mulle_default_allocator;
-
    buffer = mulle_allocator_malloc( allocator, sizeof( struct mulle_buffer));
    mulle_buffer_init( buffer, allocator);
    return( buffer);